document NewPlayer
