+++ new-player/src/main/java/net/newpipe/newplayer/data/PlayMode.kt
@@ -1,0 +1,33 @@
+/* NewPlayer
+ *
+ * @author Christian Schabesberger
+ *
+ * Copyright (C) NewPipe e.V. 2024 <code(at)newpipe-ev.de>
+ *
+ * NewPlayer is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * NewPlayer is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with NewPlayer.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+
+package net.newpipe.newplayer.data
+
+enum class PlayMode {
+    IDLE,
+    EMBEDDED_VIDEO,
+    FULLSCREEN_VIDEO,
+    PIP,
+    BACKGROUND_VIDEO,
+    BACKGROUND_AUDIO,
+    FULLSCREEN_AUDIO,
+    EMBEDDED_AUDIO
+}
